Extract game core (#29)
