+++ src/core/animation.hpp
