+++ src/core/util.hpp
